pyproject.toml
[tool.poetry]
name = "skg"
version = "0.1.0"
description = ""
authors = ["Pietro Lesci"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
# openai = "^0.23.0"
# sqlparse = "^0.4.2"
torch = "^1.12.1"
transformers = "^4.23.1"
sentence-transformers = "^2.2.2"
ipykernel = "^6.15.3"
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
nbqa = "^1.5.1"
Glances = {extras = ["all"], version = "^3.2.7"}
toml = "^0.10.2"
ipywidgets = "^7.7"
datasets = "~2.5"
numpy = "^1.23.3"
pandas = "^1.5.0"
nltk = "^3.7"
hydra-core = "^1.2.0"
hydra-colorlog = "^1.2.0"
pytorch-lightning = "^1.7.7"
torchmetrics = "^0.10.0"
rich = "^12.6.0"
wandb = "^0.13.3"
srsly = "^2.4.4"
dvc = "^2.29.0"
torch-optimizer = "^0.3.0"
dvc-gdrive = "^2.19.0"
deepdiff = "^5.8.1"
pytest = "^7.1.3"
tqdm = "^4.64.1"
matplotlib = "^3.6.1"
fuzzywuzzy = "^0.18.0"
parse = "^1.19.0"
scikit-learn = "^1.1.2"
openpyxl = "^3.0.10"
polars = "^0.14.25"
swifter = {extras = ["groupby"], version = "^1.3.4"}
more-itertools = "^9.0.0"
deepspeed = "^0.7.4"
seaborn = "^0.12.1"
numba = "^0.56.4"
bitsandbytes = "^0.35.4"
accelerate = "^0.14.0"
python-Levenshtein = "^0.20.8"
polyfuzz = "^0.4.0"
fairscale = "^0.4.12"
sagemaker = "^2.117.0"

[tool.black]
line-length = 120
skip-string-normalization = false
target-version = ['py37', 'py38', 'py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[experiments.openai]
api_key = "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"

[experiments.directories]
output_dir = "/home/lescipi/IC-DST/outputs"
